+++ person.rb
@@ -1,4 +1,6 @@
 require_relative 'corrector'
+#require_relative './teacher'
+#require_relative './student'
 
 class Person
   def initialize(age, name = 'unknown', parent_permission: true)
@@ -8,7 +10,6 @@ class Person
     @permission = parent_permission
     @correct = Corrector.new
     @rental = []
-    @people = []
   end
 
   attr_reader :id, :name, :age, :permission, :rental
@@ -35,49 +36,6 @@ class Person
 
   def add_rental(book, data)
     @rental.push(Rental.new(data, self, book))
-  end
-
-  def list_all_people
-    @people.each { |n| puts n.to_string }
-  end
-
-  def create_a_student
-    default_classroom = Classroom.new('default')
-    puts 'Age: '
-    age = gets.chomp
-    puts 'Name: '
-    name = gets.chomp
-    puts 'Has parent permission? [Y/N]: '
-    permission = gets.chomp
-    student = if permission.upcase == 'Y'
-                Student.new(age, default_classroom, name, parent_permission: true)
-              else
-                Student.new(age, default_classroom, name, parent_permission: false)
-              end
-    @people.push(student)
-  end
-  
-  def create_a_teacher
-    puts 'Age: '
-    age = gets.chomp
-    puts 'Name: '
-    name = gets.chomp
-    puts 'Specilaization'
-    specialization = gets.chomp
-    teacher = Teacher.new(age, specialization, name)
-    @people.push(teacher)
-  end
-  
-  def create_a_person
-    puts 'Do you want to create a Student (1) or a Teacher (2)? [Input the number]: '
-    var = gets.chomp
-    case var
-    when '1'
-      create_a_student
-    when '2'
-      create_a_teacher
-    end
-    puts 'Person created Succesfully'
   end
 
   private
